requirements.txt
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
billiard==3.6.4.0
bleach==5.0.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.0.0
dj-rest-auth==3.0.0
Django==4.2.2
django-admin-sortable2==2.1.8
django-allauth==0.54.0
django-angular==2.3
django-appconf==1.0.5
django-classy-tags==3.0.1
django-cms==3.11.3
django-cors-headers==3.13.0
django-countries==7.3.2
django-entangled==0.5.4
django-filer==2.2.4
django-formtools==2.4
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-ipware==5.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==6.1.0
django-polymorphic==3.1.0
django-post-office==3.6.3
django-sekizai==3.0.1
django-select2==7.10.0
django-theme-pixel==1.0.3
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap5==0.1.0
djangocms-cascade==2.0.1
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.1.1
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.2
djangocms-video==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.1
easy-thumbnails==2.8.5
google-api-core==2.8.2
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
humanize==4.4.0
idna==3.3
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.16.0
kombu==5.2.4
lxml==4.9.2
mysqlclient==2.1.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.47
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.21.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
reportlab==3.6.12
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
twilio==7.9.3
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==4.3
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
django_extensions
django-crispy-forms
crispy-bootstrap5
djangocms-blog